+++ runelite-api/src/main/java/net/runelite/api/Projection.java
@@ -26,5 +26,5 @@ package net.runelite.api;
 
 public interface Projection
 {
-	float[] project(int x, int y, int z);
+	float[] project(float x, float y, float z);
 }
